Fix percentile reduction
