+++ nncf/common/tensor_statistics/collectors.py
@@ -12,48 +12,20 @@
 from abc import ABC
 from abc import abstractmethod
 from collections import deque
-from copy import deepcopy
-from typing import Callable, Deque, Dict, List, Optional, Tuple, Union
+from typing import Deque, Dict, List, Optional, Tuple
 
-import numpy as np
-
-from nncf.common.quantization.structs import QuantizerScaleShape
 from nncf.common.tensor import NNCFTensor
+from nncf.common.tensor_statistics.reduction import ReductionAxes
 from nncf.common.tensor_statistics.reduction import get_per_channel_history
-from nncf.common.tensor_statistics.reduction import percentile_reduce_like
+from nncf.common.tensor_statistics.reduction import get_reduction_shape_from_sample_shape
+from nncf.common.tensor_statistics.reduction import is_reduce_to_scalar
+from nncf.common.tensor_statistics.reduction import percentile_reduce
 from nncf.common.tensor_statistics.statistics import MeanTensorStatistic
 from nncf.common.tensor_statistics.statistics import MedianMADTensorStatistic
 from nncf.common.tensor_statistics.statistics import MinMaxTensorStatistic
 from nncf.common.tensor_statistics.statistics import PercentileTensorStatistic
 from nncf.common.tensor_statistics.statistics import RawTensorStatistic
 from nncf.common.tensor_statistics.statistics import TensorStatistic
-
-ReductionAxes = Tuple[int]
-ReductionShape = Tuple[int]
-
-REDUCE_TO_SCALAR_REDUCTION_SHAPE = (-1,)
-
-
-def get_reduction_axes_from_scale_shape(scale_shape: QuantizerScaleShape) -> ReductionAxes:
-    if scale_shape.is_per_tensor():
-        return REDUCE_TO_SCALAR_REDUCTION_SHAPE
-    return tuple(i for i, dim in enumerate(scale_shape.shape) if dim == 1)
-
-
-def is_reduce_to_scalar(reduction_axes: ReductionAxes) -> bool:
-    return reduction_axes == REDUCE_TO_SCALAR_REDUCTION_SHAPE
-
-
-def get_reduction_shape_from_sample_shape(sample_shape: List[int], reduction_axes: ReductionAxes) -> ReductionShape:
-    if is_reduce_to_scalar(reduction_axes):
-        return (1,)
-    reduced_shape = deepcopy(list(sample_shape))
-    for ax in reduction_axes:
-        reduced_shape[ax] = 1
-    return tuple(reduced_shape)
-
-
-MaskedReduceFN = Callable[[NNCFTensor, Union[int, tuple, list], NNCFTensor, bool], NNCFTensor]
 
 
 class TensorStatisticCollectorBase(ABC):
@@ -449,7 +421,7 @@ class MeanPercentileStatisticCollector(OfflineTensorStatisticCollector):
 
     def _register_input(self, x: NNCFTensor):
         for pct, values in self._all_pct_values.items():
-            pct_vals = percentile_reduce_like(x, self._reduction_axes, pct)
+            pct_vals = percentile_reduce(x, self._reduction_axes, pct)
             values.append(pct_vals)
 
     def _get_statistics(self) -> PercentileTensorStatistic:
